Import pgsql-cluster-manager pkg
We're decommissioning pgsql-cluster-manager in favour of this project.
Instead of importing the entire pgcm pkg, we should move the code here
to ensure we continue to run the package tests on each build.

This commit imports the code almost wholesale, with a couple of
modifications to suit our new use case.