Run tests from host machine
This change exposes the PgBouncer ports on the host machine in order for
the acceptance test suite to be run locally. This unifies the
perspective of the acceptance test suite with a developer who is
exploring the cluster via the exposed docker-compose ports, and enables
a naive `go run cmd/stolon-pgbouncer-acceptance/main.go` to succeed
provided the docker-compose setup is already up.