Adjust stolon interval/timeout & specify dev image
Lower the stolon failInterval so we don't wait too long in or test for
the async keeper that we pause to be marked as failed. Also lower the
requestTimeout as all requests are local in the tests.

We also specify the stolon-development image instead of using a build
context for extra clarity. With a number of moving parts we found it
easier to grok. Feel free to revert if it causes a workflow issue when
developing.