failover --health-check-only, check PgBouncer
Provide the option to just health check the failover endpoints, useful
if you want to validate everything is responsive in a test environment
etc.

Also check the PgBouncer health when running health checks. This makes
it easy to verify everything is ready before committing to a failover,
potentially pausing other working PgBouncers when you'll inevitably fail
due to a bad node.