Reorder acceptance tests to be more RSpec
This commit re-arranges the acceptance test suite to be more like RSpec
by using the ginkgo JustBeforeEach idiom, emulating the RSpec `let`
construct. By doing this we considerably DRY up the tests and hopefully
make our test names better represent what they intend to check.

This change was made in response to adding tests for the health check,
and contains those tests as well. We remove the test checking
asynchronous standby failure as it wasn't quite testing what we wanted
it to, but have ticketed up the work to amend that (and provide the
functionality we hoped we were testing for).