rm -rf pkg/acceptance/matchers
We avoid the need to implement this by adding a CheckHealthy method to
the stolon clusterdata struct which results in much less code than these
matchers. If we need to bring these back then we should consider using
the gstruct AllElements matcher to match against slices, as we may be
duplicating behaviour.