Don't require ignore_startup_parameters
[^1]: https://github.com/lib/pq/issues/475

lib/pq sends extra_float_digits as a startup parameter [^1] whenever
initialising a Postgres connection. PgBouncer doesn't recognise this
parameter and will fail whenever you send it.

To prevent this, we ensured pgsql-cluster-manager would ban PgBouncer
configuration that didn't explicitly ignore this parameter. This is
likely irrelevant after we moved to pgx though, so we should remove this
constraint.