Expose keeper as label
Prometheus discourages this type of behaviour, where you add dynamic
information into labels. In this case we'll have very few host values
that will appear here, so it doesn't matter so much if we do this.

We'll use this label to provide a recognisable textual host value in our
diagnostic dashboards, allowing us to easily list which pods are
connected to which host.