stolon_pgbouncer_cluster_identifier = 1
We label this metric with the relevant information, so don't need to set
the value to the hash of our cluster. I had initially suspected doing so
might be useful, but I've ended up multiplying this value by 0 and
adding 1 in every case I've used it, implying it's not useful at all.