Boot Grafana and Prometheus in docker-compose
This commit adds Grafana and Prometheus containers to the docker-compose
stack, enabling developers to locally verify their metrics, alerts and
dashboards. Some of the alerts are a bit suspect right now but this is
the first step in making them good.

Also ensure Circle doesn't boot these containers when running acceptance
tests. While it doesn't cause the test to fail, there's no point
starting prometheus and friends in an integration testing environment.