Expose and use keeper metrics
[^1]: https://github.com/gocardless/stolon/pull/1

This commit is associated with an open PR [^1] to stolon that adds
Prometheus metrics to the keeper. The changes here include adding a
keeper dashboard that can visualise the keeper statuses and a couple of
essential alerts for keeper health.

We update the playground environment so developers can explore these
metrics and make use of the dashboard. This includes a new docker image,
which has been pushed to Docker hub.