Retry watch indefinitely
Instead of terminating whenever we receive an error from the etcd
watcher, instead retry indefinitely on the assumption that this will
eventually succeed. This prevents us going into a termination state from
a recoverable etcd error, and we can rely on the downstream stolon proxy
to ensure we never connect to a bad (non-primary) node.

We also make our watcher more fussy, by rejecting any watch streams that
have ended up coming from a non-leader etcd node. This should ensure we
don't ever get stuck speaking to an etcd node that is partitioned from
the leader and not try to reconnect to a more up-to-date member.