Test for pauser boot-resume
We should be robust to our pauser dying mid-migration. This means when
we restart, the pauser should issue a resume to PgBouncer to remove any
pause instructions that were mid-operation in a previous process life.

Adding this test as a first commit to show it going red, then following
with a commit that adds this behaviour.