Install pgbouncer_exporter to development docker
Configure the stolon-development image to install the GC fork of
pgbouncer_exporter. Prior to this change, the stolon-pgbouncer docker
container (named pgbouncer) would run stolon-pgbouncer as pid 1.

To support multiple binaries within the same container, we have to
create a new supervisord config that manages what used to be pid 1 along
with the pgbouncer_exporter.

Finally, we expose the conventional pgbouncer_exporter port (9127) on
the development container.