Consolidate Prometheus metric labels
GoCardless use the instance label to mean the human readable identifier
of the scraped target. This commit modifies the development Prometheus
to have instance set to the hostname of the container, along with a port
label from deconstructing __address__.

The alerts are then modified to use instance instead of pod_name, which
was never going to work in our local development version anyway. This
will be portable for most people using Prometheus.