Grafana dashboards support new labels & PgBouncer
Now we have a better labelling system (with instance as a human readable
hostname) dashboards need updating to support it.

We also have PgBouncer metrics now, and should use these to visualise
activity associated with keeper and supervised PgBouncers. We filter the
relevant PgBouncer metrics by joining onto the shutdown metric which has
different names depending on running in the keeper or stolon-pgbouncer.

There's one more fancy addition, which is to help operators manage the
rolling of stolon-pgbouncer deployments within kubernetes. If the
metrics have an associated pod_template_hash, then Grafana will create a
panel per active version of stolon-pgbouncer available for the selected
cluster identifier. It's then possible to watch as the old version
slowly drains of traffic and see as new queries hit the new rollout.