Increase default TerminationGracePeriod
The default TerminationGracePeriod of 5s was not long enough for our
stolon-pgbouncer deployment in kubernetes. When the pod would begin
termination, it would take up to 8 seconds for no new connectins to be
sent to it. This would lead to a number of failed connections after the
TerminationGracePeriod had been reached.

We now set this to 15s as the default which should have no negative
impact but have the benefit of others not having to diagnose and fix
the same issue we have been experiencing.