Report pause duration after failover
It's useful to know how long traffic was paused during failover.
Especially now we wait for the sleepInterval to be loaded, it's not
obvious how long traffic was actually impacted by watching the output of
stolon-pgbouncer failover.

This change adds a log entry to report the time between attempting to
pause PgBouncers and them successfully resuming.