rm -rf vendor
Go modules package mirror looks way more reliable than direct to Github,
and the community is a lot more aware of how people are consuming their
modules than they used to be. This decreases the risk that a package we
depend on might get yanked, so it makes more sense for us to drop the
vendor directory and be able to fully benefit from dependabot updating
our deps.