Add stolon_store_certificate_expiry_seconds metric
Adds stolon_store_certificate_expiry_seconds metric which stores the
expiry of the etcd store certificate that stolon-pgbouncer is using to
connect to etcd. We also add an alert to let us know when expiration is
approaching so we can update it before its expiry prevents
stolon-pgbouncer connecting to etcd.