Bump gocardless-v0.17.0+ and enable pg_rewind
While we had set `usePgrewind` to true in the cluster spec in the
playground we were never using pg_rewind as we we using the `trust` auth
method for the posgres su user.

Setting is back to the default of `md5` as well as supplying a pg su
user and password to the keeper is required for pg_rewind.

We also bump the stolon fork to the latest sha in gocardless-v0.17.0+
which has some pg_rewind specific changes in it that we wanted to test
hence the enabling of pg_rewind above.