Add healthchecks and build contexts to the docker-compose config
This change adds some healthchecks, and the tracking of said
healthchecks for dependencies to start in a more consistent manner.

Also included in this change are `build` rules for when the provided
images aren't already available on the machine, falling back to
building them each time. This additionally provides the ability to run
`docker compose build` to rebuild relevant images.